requirements.txt
insightface
numpy
onnxruntime-gpu
opencv-python
wxPython
